refactor(catalog): извлечён CatalogResource + системный каталог на нём (#210 ось видимости, PR1)
Первый шаг распространения list-detail на ось scope (разбор с Дизайнером+Архитектором+другом).
Извлечён `CatalogResource({scope,scopeId,allDocTypes})` — богатый браузер каталога общих данных
(поиск + фильтр по типу + группы + add/edit/delete), scope-параметризованный, БЕЗ чипа области
(scope выражается положением, не чипом). `SystemCommonDataPage` сведён к тонкой обёртке над ним.
Единый компонент — задел под detail-панели scope-страниц (стройка/раздел/комплект) в следующих PR.
